requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0 